Fix game crash on level selection and address build warnings
The game was crashing when a level was selected from the menu. This was caused by two issues:
1. A missing implementation for loading and rendering level data.
2. Functions in `Player/BodyPart.cpp` not returning a value on all control paths, leading to undefined behavior when the player is spawned.

This change introduces the necessary logic to define, load, and render game levels, and it fixes the critical warnings in the player logic.

- Defined a data structure for level layouts in `LevelData.h`.
- Populated `LevelModel` with data for two levels.
- Implemented level loading logic in `LevelModel` and `LevelController`.
- Updated `LevelService` to trigger level loading when a level is selected.
- Updated `LevelView` to render the level layout, including obstacles.
- Added a defensive check in `LevelView` to prevent a crash if the obstacle texture fails to load.
- Fixed functions in `Player/BodyPart.cpp` to ensure they return a value on all control paths.
- Addressed various C4244 conversion warnings with explicit `static_cast`s.

+++ Linked-List-Snake/source/Level/LevelController.cpp
@@ -20,7 +20,7 @@ namespace Level
 	void LevelController::initialize()
 	{
 		level_view->initialize();
-		level_model->initialize(level_view->getGridWidth(), level_view->getGridHeight());
+		level_model->initialize(static_cast<int>(level_view->getGridWidth()), static_cast<int>(level_view->getGridHeight()));
 	}
 
 	void LevelController::update()
